+++ example/websocket/client/coro/websocket_client_coro.cpp
@@ -15,9 +15,7 @@
 
 #include <boost/beast/core.hpp>
 #include <boost/beast/websocket.hpp>
-#include <boost/asio/connect.hpp>
 #include <boost/asio/spawn.hpp>
-#include <boost/asio/ip/tcp.hpp>
 #include <cstdlib>
 #include <functional>
 #include <iostream>
@@ -51,15 +49,19 @@ do_session(
 
     // These objects perform our I/O
     tcp::resolver resolver{ioc};
-    websocket::stream<tcp::socket> ws{ioc};
+    websocket::stream<
+        beast::tcp_stream<net::io_context::executor_type>> ws(ioc);
 
     // Look up the domain name
     auto const results = resolver.async_resolve(host, port, yield[ec]);
     if(ec)
         return fail(ec, "resolve");
 
+    // Set a timeout on the operation
+    beast::get_lowest_layer(ws).expires_after(std::chrono::seconds(30));
+
     // Make the connection on the IP address we get from a lookup
-    net::async_connect(ws.next_layer(), results.begin(), results.end(), yield[ec]);
+    beast::async_connect(ws.next_layer(), results, yield[ec]);
     if(ec)
         return fail(ec, "connect");